Remove jsonutil module
