+++ internal/log_analysis/awsglue/quirks.go
@@ -24,7 +24,7 @@ import (
 
 	jsoniter "github.com/json-iterator/go"
 
-	"github.com/panther-labs/panther/internal/log_analysis/log_processor/jsonutil"
+	"github.com/panther-labs/panther/internal/log_analysis/log_processor/pantherlog/renamefields"
 	"github.com/panther-labs/panther/internal/log_analysis/log_processor/pantherlog/tcodec"
 )
 
@@ -73,7 +73,7 @@ func (*timestampEncoder) EncodeTime(tm time.Time, stream *jsoniter.Stream) {
 }
 
 func RegisterExtensions(api jsoniter.API) jsoniter.API {
-	api.RegisterExtension(jsonutil.NewEncoderNamingStrategy(RewriteFieldName))
+	api.RegisterExtension(renamefields.New(RewriteFieldName))
 	api.RegisterExtension(tcodec.NewExtension(tcodec.Config{
 		// Force all timestamps to be awsglue format and UTC. This is needed to be able to write
 		DefaultCodec: tcodec.Join(nil, NewTimestampEncoder()),
