+++ internal/log_analysis/log_processor/pantherlog/null/null.go
@@ -91,3 +91,13 @@ func ValidateNullType(val reflect.Value) interface{} {
 	}
 	return nil
 }
+
+func unquoteJSON(data []byte) []byte {
+	if len(data) > 1 && data[0] == '"' {
+		data = data[1:]
+		if n := len(data) - 1; 0 <= n && n < len(data) && data[n] == '"' {
+			return data[:n]
+		}
+	}
+	return data
+}
