+++ internal/log_analysis/log_processor/pantherlog/null/uint64.go
@@ -24,8 +24,6 @@ import (
 	"unsafe"
 
 	jsoniter "github.com/json-iterator/go"
-
-	"github.com/panther-labs/panther/internal/log_analysis/log_processor/jsonutil"
 )
 
 // Uint64 represents a nullable uint64 value.
@@ -57,7 +55,7 @@ func (u *Uint64) UnmarshalJSON(data []byte) error {
 		return nil
 	}
 	// Handle both string and number input
-	data = jsonutil.UnquoteJSON(data)
+	data = unquoteJSON(data)
 	// Empty string is considered the same as `null` input
 	if len(data) == 0 {
 		*u = Uint64{}
